Added port enumeration and big refactor
