+++ src/ssh_manager.py
@@ -1,5 +1,4 @@
 import logging
-import paramiko
 from fabric import Connection, Config
 
 
@@ -18,7 +17,7 @@ class SSHManager:
                 self.hostname, user=self.username, config=self.config, port=22,
                 connect_kwargs={'password': self.password})
 
-            #self.setup_environment()
+            self.setup_environment()
         except Exception as e:
             print(f"Unable to connect to {self.hostname}: {e}")
             raise ConnectionError(e)
@@ -35,20 +34,8 @@ class SSHManager:
         if self.client is None:
             raise ValueError("Connection not established.")
 
-        command = "sudo -S -p '' %s" % command
-        logging.info(f"Executing command: {command}")
-
-        stdin, stdout, stderr = self.client.exec_command(command=command)
-        stdin.write(self.password + "\n")
-        stdin.flush()
-
-        stdoutput = [line for line in stdout]
-        stderroutput = [line for line in stderr]
-        for output in stdoutput:
-            logging.info("Job: %s" % (output.strip()))
-
-        return stdoutput, stderroutput
+        return self.client.sudo(command)
 
     def setup_environment(self):
-        self.execute_command_as_root('apt update')
+        self.execute_command_as_root('apt update -y')
         self.execute_command_as_root('apt install -y curl zip')
